Fix collect_forces to use file handle after I broke it.
When I modified the collect_forces function in an earlier commit
I assumed f was the text of the file but it is actually a file
handle. This explains why the loop with the hook was written as
it was. Previously, I assumed the loop wasn't doing anything
but that was incorrect. Anyway, I've changed it now so it should
be compatible with all the different interfaces again.